+++ notebooks/shared.py
@@ -1,20 +1,21 @@
+import os
 from pathlib import Path
+from zipfile import ZipFile
 
 import geopandas as gpd
 import xarray as xr
-from distributed import Client
-from distributed.diagnostics.plugin import UploadDirectory
+from dask_gateway import GatewayCluster
 
 from pyCIAM import __file__
 
 DIR_SCRATCH = Path("/tmp/ciam-scratch")
 
-SLIIDERS_VERS = "v1.1"
-RES_VERS = "v1.1"
+SLIIDERS_VERS = "v1.2"
+RES_VERS = "v1.2"
 
 # Cloud Storage tools (will work with local storage as well but may need to be specifiec
 # for cloud buckets
-STORAGE_OPTIONS = {}
+STORAGE_OPTIONS = None
 
 
 def _to_fuse(path):
@@ -25,9 +26,10 @@ def _to_fuse(path):
 QUANTILES = [0.17, 0.5, 0.83]
 
 # Output dataset attrs
-HISTORY = """version 1.1: Version associated with Depsky et al. 2023"""
+HISTORY = """version 1.1: Version associated with Depsky et al. 2023.
+version 1.2: Updated to use SLIIDERS v1.2 as input."""
 AUTHOR = "Ian Bolliger"
-CONTACT = "ian.bolliger@blackrock.com"
+CONTACT = "ian@reask.earth"
 
 # AR5 SLR projections info
 LOCALIZESL_COREFILES = {
@@ -54,9 +56,9 @@ DIR_RES = DIR_HOME / f"results-{RES_VERS}"
 ##################
 
 # NECESSARY FOR EXAMPLE
-PATH_PARAMS = Path.home() / "git-repos/pyciam/params.json"
+PATH_PARAMS = Path.home() / "git-repos/pyCIAM/params.json"
 
-PATH_PARAMS_DIAZ = Path.home() / "git-repos/pyciam/params_diaz.json"
+PATH_PARAMS_DIAZ = Path.home() / "git-repos/pyCIAM/params_diaz.json"
 
 ##################
 # SOCIOECON INPUTS
@@ -67,8 +69,9 @@ PATH_PARAMS_DIAZ = Path.home() / "git-repos/pyciam/params_diaz.json"
 # NECESSARY FOR EXAMPLE
 PATH_SLIIDERS = DIR_RAW / f"sliiders-{SLIIDERS_VERS}.zarr"
 
-# NECESSARY FOR EXAMPLE
-PATH_SLIIDERS_SEG = DIR_INT / f"sliiders-{SLIIDERS_VERS}-seg.zarr"
+PATH_SLIIDERS_SEG = PATH_SLIIDERS.parent / (
+    PATH_SLIIDERS.stem + "-seg" + PATH_SLIIDERS.suffix
+)
 
 # Diaz
 PATH_DIAZ_INPUTS_RAW = DIR_RAW / "diaz2016_inputs_raw.zarr"
@@ -132,7 +135,7 @@ PATH_DIAZ_RES = DIR_RES / "diaz2016_outputs.zarr"
 # FILES FOR PLOTTING RESULTS
 ############################
 PATH_MOVEFACTOR_DATA = DIR_RES / "suboptimal_capital_by_movefactor.zarr"
-PATH_SLIIDERS_INCOME_INTERMEDIATE_FILE = DIR_RAW / "ypk_2000_2100_20221122.zarr"
+PATH_SLIIDERS_INCOME_INTERMEDIATE_FILE = DIR_RAW / "ypk_2000_2100_20240222.zarr"
 
 DIR_FIGS = Path("/home/jovyan/ciam-figures")
 DIR_SHP = DIR_RAW / "shapefiles"
@@ -155,48 +158,48 @@ for p in [
     p.mkdir(exist_ok=True, parents=True)
 
 
-def upload_pyciam(client, restart_client=True):
-    """Upload a local package to Dask Workers. After calling this function, the package
-    contained at ``pkg_dir`` will be available on all workers in your Dask cluster,
-    including those that are instantiated afterward. This package will take priority
-    over any existing packages of the same name. This is a useful tool for working with
-    remote dask clusters (e.g. via Dask Gateway) but is not needed for local clusters.
-    If you wish to use this, you should call this function from within
-    `start_dask_cluster`.
+def _zipdir(
+    path,
+    zip_filename,
+    skip_files=(
+        ".git",
+        ".github",
+        ".pytest_cache",
+        "tests",
+        "docs",
+        "deploy",
+        "notebooks",
+        ".ipynb_checkpoints",
+        "__pycache__",
+        ".coverage",
+        "dockerignore",
+        ".gitignore",
+        ".gitlab-ci.yml",
+        ".gitmodules",
+        "pyclaw.log",
+        "run_tests.sh",
+    ),
+):
 
-    Parameters
-    ----------
-    client : :py:class:`distributed.Client`
-        The client object associated with your Dask cluster's scheduler.
-    pkg_dir : str or Path-like
-        Path to the package you wish to zip and upload
-    **kwargs
-        Passed directly to :py:class:`distributed.diagnostics.plugin.UploadDirectory`
-    """
-    client.register_worker_plugin(
-        UploadDirectory(
-            Path(__file__).parents[1],
-            update_path=True,
-            restart=restart_client,
-            skip_words=(
-                ".git",
-                ".github",
-                ".pytest_cache",
-                "tests",
-                "docs",
-                "deploy",
-                "notebooks",
-                ".ipynb_checkpoints",
-                "__pycache__",
-                ".coverage",
-                "dockerignore",
-                ".gitignore",
-                ".gitlab-ci.yml",
-                ".gitmodules",
-                "pyclaw.log",
-            ),
-        )
-    )
+    with ZipFile(zip_filename, "w") as ziph:
+        for root, dirs, files in os.walk(path):
+            for file in files:
+                if any([f in file.split("/") for f in skip_files]):
+                    continue
+                # Create a relative path for files to preserve the directory structure
+                # within the ZIP archive. This relative path is based on the directory
+                # being zipped, so files are stored in the same structure.
+                relative_path = os.path.relpath(
+                    os.path.join(root, file), os.path.join(path, "..")
+                )
+                ziph.write(os.path.join(root, file), arcname=relative_path)
+
+
+def upload_pyciam(client):
+    package_dir = Path(__file__).parent
+    zip_filename = "/tmp/pyCIAM.zip"  # Output ZIP file name
+    _zipdir(package_dir, zip_filename)
+    client.upload_file(zip_filename)
 
 
 def save(obj, path, *args, **kwargs):
@@ -223,6 +226,12 @@ def open_dataset(path, **kwargs):
     return xr.open_dataset(_path, **kwargs)
 
 
+def save_dataset(ds, path, **kwargs):
+    _path = str(_to_fuse(path))
+    _generate_parent_fuse_dirs(_path)
+    return ds.to_netcdf(_path, **kwargs)
+
+
 def open_dataarray(path, **kwargs):
     _path = str(_to_fuse(path))
     _generate_parent_fuse_dirs(_path)
@@ -235,7 +244,8 @@ def read_shapefile(path, **kwargs):
     return gpd.read_file(_path, **kwargs)
 
 
-def start_dask_cluster(**kwargs):
-    client = Client(**kwargs)
-    print(client.dashboard_link)
-    return client
+def start_dask_cluster(profile="micro", **kwargs):
+    cluster = GatewayCluster(profile=profile, **kwargs)
+    client = cluster.get_client()
+    upload_pyciam(client)
+    return client, cluster