+++ HISTORY.rst
@@ -1,10 +1,16 @@
 History
 =======
 
-v1.1.3 (unreleased)
--------------------
+unreleased
+----------
+* Update GitHub actions versions
+
+v1.2.0
+------
 * Point `data-acquisition.ipynb` to updated Zenodo deposit that fixes the dtype of `subsets` variable in `diaz2016_inputs_raw.zarr.zip` to be bool rather than int8
 * Variable name bugfix in `data-acquisition.ipynb`
+* Add netcdf versions of SLIIDERS and the pyCIAM results to `upload-zenodo.ipynb`
+* Update results in Zenodo record to use SLIIDERS v1.2
   
 v1.1.2
 ------